fix(ldap2mo): [#59851] fix org-unit creation
This commit fixes an issue with the `create_org_unit` method on the
`LdapConverter` class, unlike what the name says it does in fact not
create a single org-unit, but rather the entire ancestor line from a
desired org-unit to an org-unit in the existing tree within OS2mo.

The issue with the code was in the dataloader `create` method, which
used to handle org-unit creation, but had its behavior changed such that
it simply raised an exception when trying to create org-units.

The issue was introduced in version 10.25.0 of the integration with !708
(cad9de69), which split the `create_object` method into type specific
handler methods in preparation for the elimination of the model-client
and its replacement with GraphQL.
However during this split the creation of org-units within the program
was missed.

This commit fixes the issue by introducing a `create_org_unit` function
on the `DataLoader` class, which as the name implies, creates a single
org-unit within OS2mo.

The issue has not resulted in any erroneous writes, thus data integrity
should be entirely unaffected. The change may however have resulted in
blocked synchronizations, such that some entities in MO have not been
correctly updated. This should have been a non-issue as the AMQP messages
should have been piling up, both serving to alert us to the issue, and
also ensuring that the backlog of broken synchronizations could be
rerun.

It is unfortunate that our static analysis typing tool mypy was not
unable to detect this discrepancy, seeing as the called method only
accepts a subset of types, however due to lax typing this was not
possible. A future commit will work to tighten up this, such that our
static analysis tool will detect similar problems in the future.

The issue was a subtle one that was fairly not caught in the review
process, however it should have (and likely would have) been caught if
we had had integration tests or less mocked unit-tests.
This commit does not introduce such tests, nor does it demockify the
current unit-tests, rather a future commit will address this need.