feat(ldap2mo): eliminate dns_to_ignore
This commit eliminates `dns_to_ignore` from `SyncTool`, the reasoning
behind its introduction is still unclear, it could be meant as either an
optimization / deduplication or as a tool to achieve correctness, but as
it stands it fails as both.

For the purpose of optimization or deduplication it is harmful as it
allows the first synchronization, but blocks all subsequent updates,
thereby losing potentially important updates. A deduplication mechanism
has to rather block everything but the last update to ensure that no
updates are lost.

For the purpose of correctness, it works to avoid the integration
triggering it self in an infinite loop, however such a situation should
only occur if the integration does not have a well-defined end state,
which is then the core issue that needs to be resolved.

As the integration now has a well-defined end state, it is our belief
that `dns_to_ignore` can be removed without any ill effects, but rather
that its removal will increase correctness by ensuring that updates are
no longer lost.

If deduplication is desired for the purpose of optimization later, it
should be implemented as a feature in the upcoming AMQP-to-HTTP bridge.