refactor: [#59500] remove exception from extract_current_or_latest_validity
This commit refactors the `extract_current_or_latest_validity` function
such that it no longer raises an exception when no result is found, but
rather return `None` when this happens.

We prefer returning `None` as it is typed and forces the caller to deal
with situation rather than silently having the exception pass through
the call-stack.

Note that the method still can raise a `ValueError` if multiple current
validities are found, this should probably be resolved in the future.