ci: [#59372] upgrade integration test template
This commit upgrades the local copy of the OS2mo integration test CI
template, as required for newer versions of FastRAMQPI.

The local copy unfortunately has to be maintained as the integration
needs to run additional services during integration test (an LDAP server),
and thus needs an extra layer of inheritance in the CI templates, which
is not supported by Gitlab. See the commit introducing the template for
further details (5fa2a601).