build(deps): [#56970] upgrade FastAPI
This commit upgrades FastAPI to the newest version, adding an odd piece
of code to our exception handling to enable the upgrade.

The odd piece of code is needed because of FastAPI 0.108.0, which
upgrades Starlette to a newer version than 0.29.0, which introduces a
custom `__str__` method to `HTTPException`, which breaks our exception
interface. See: https://github.com/encode/starlette/pull/2181 for details.

The odd piece of code simply circumvents the `__str__` implementation,
and uses the old `Exception.__str__` behavior instead, ensuring
backwards compatibility and thus ensuring that the upgrade is not a
breaking change.

The newer version of FastAPI also required an bugfix upgrade of RAMQP,
to version 9.1.1 as prior versions of RAMQP does not work with newer
versions of FastAPI.