feat: [#59412] fetch and emit LDAP entry UUIDs directly
This commit simplifies the LDAP event generation code by fetching the
LDAP entry UUIDs directly, instead of fetching the DNs from LDAP and
then converting those to LDAP entry UUIDs.

This should increase the performance of the pollers, while also guarding
against some issues where an DN is found with the first lookup, but
missing for the conversion due to the entity being moved, renamed or
similar.

Additionally this commit also introduces a poor deduplication by storing
the changed UUIDs in a set before emitting events, in the future however
deduplication will be done on the AMQP2HTTP bridge, as messages may
already exist there with the same UUIDs.