fix: [#59412] fix datetime_to_ldap_timestamp
This commit fixes an issue with the `datetime_to_ldap_timestamp` function,
where it would produce incorrect LDAP timestamps, not in accordance with
LDAP ASN.1 GeneralizedTime format (OID: 1.3.6.1.4.1.1466.115.121.1.24),
i.e. not an LDAP encoded ISO 8601 date.

The issue was in the part of the code that handled the fractional second
part of the  produced timestamp, the code would simply append the number
of milliseconds at the end of the timestamp, for instance, the timestamp
for `2020-01-01T12:00:00` would be encoded as: `20200101120000`.
Adding 2000 milliseconds to this would be encoded as a `.2` suffix, while
adding 20000 milliseconds to it would be encoded as a `.20` suffix, etc.
Thus adding further milliseconds would add precision, not time.

The issue was shipped in version 1.0.0 of the integration and was merged
before every incremental change to the integration was releaed with !101
(51e8391f). The commit introducing the change did not have any comments
regarding the issue in the review, and as such it seems to have been
missed.

This commit fixes the issue by generating the timestamp entirely using
the `strftime` standard library method.

The issue has not resulted in any erroneous writes, thus data integrity
should be entirely unaffected. It is however very likely that the issue
has resulted in missing synchronizations, as the generated timestamps
likely would have had gaps.