refactor(ldap_event_generator): [#59412] pass search_base instead of search_parameters
This commit changes how the entire callstack from `setup_poller` all the
way to `_poll` receives LDAP search parameters.

Prior to this change the search parameters were constructed inside the
`LDAPEventGenerator` and passed all the way down the stack to be modified
in-place inside `_poll`.

Now the only argument send from `LDAPEventGenerator` is the `search_base`
parameter which is then used to construct the `search_parameters`
dictionary inside of `_poll` avoiding in-place modification and ensuring
that a new clean object is used for each invocation.

It also simplifies the callstack and makes it easier to reason about what
is being send based upon the types and names of the parameters.