refactor(ldap_event_generator): [#59412] remove init_search_time argument
This commit removes the start-time parameter of the poller from inside
the `LDAPEventGenerator` to inside the `_poller` method, as it is an
equivalent transformation, and as it prepares the code nicely for the
introduction of our rundb in a future merge request.

It also simplifies the callstack as one less parameter is being passed
around.