refactor(ldap_event_generator): [#59412] let _poll caller update last_search_time
This commit removes a responsibility from the `_poll` function, namely
that it previously returned the datetime at which it had done a search,
demanding to get the same parameter back for the following call.

It is now the callers responsibility to register the time before the
function is called, and call it again in the future with this value.

The change is not exactly a refactoring, since the timing will be
slightly different, but at worst the change will result in duplicate
events, which should be handled gracefully anyway.