refactor(ldap_event_generator): [#59412] unpack _poll parameters
This commit unpacks the parameters to the `_poll` function, such that
the declaration of the function more accurately shows what dependencies
the function has, instead of simply stating that it depends on the
entire FastRAMQPI `UserContext` and everything within in.

The caller as a consequence now has to provide the specifically required
arguments.