refactor(ldap_event_generator): [#59412] reduce _poll to only finding changes
This commit reduces the responsibility of the `_poll` function, such that
it no longer is responsible for sending out AMQP events, but rather is
just responsible for contracting LDAP and generating a set of UUIDs which
may have changed.

As such it is now the responsibility of the caller to emit events if so
desired.