refactor(ldap_event_generator): [#59412] remove explicit poll_time, use settings.poll_time
This commit removes `poll_time` from the parameters to `setup_poller`,
as it was the lone argument send outside of the `user_context["settings"]`
object. It may be reintroduced in the future, if we decide to unpack the
parameters to `setup_poller` making them explicit as it was done for
`_poll`, but for now it simply aligns how parameters are handled.