feat(ldap_event_generator): [#59412] introduce last_run table
This commit introduces our first database table `last_run` which
will contain the state for the `LDAPEventGenerator` keeping track of
when it last ran, such that it can pick-up wherever it left off in case
of downtime, restarts, etc.

The commit does not actually use the model in the code, but simply
introduces it and tests that it works according to the expectations.