feat(ldap_event_generator): [#59412] introduce update_timestamp helper
This commit introduces the `update_timestamp` async context manager
helper, which will be used when we introduce the `last_run` capability
to the `LDAPEventGenerator`.

It works by reading the latest timestamp for the provided `search_base`
yielding it to the caller, only to update it with the time just before
yielding whenever the caller returns.