test: [#59412] remove pyyaml test
This commit removes a test of `pyyaml`s `safe_load` function.

It is very likely that this test used to actually test functionality
within the codebase, which has now been trivialized to just loading yaml.
However as we do not need to test `pyyaml`, since we can assume they have
tested it themselves, the test can be removed entirely.