test: [#59412] convert autouse fixture to integration-test autouse
This commit removes `autouse=True` from the `ldap_purge` fixture, in
favor of adding it automatically to tests marked with `integration_test`,
thus in effect making it an integration-test only autouse fixture.

This is done in preparation for moving the integration-test fixtures to
the common conftest file, making them available to all tests.