chore(exportchecks): [#55956] remove dead alleroed check code
This commit removes an unused customer specific export-check for Alleroed.

It is our belief that this old export check has been superseeded by
checking for the existence of a given IT-User account instead, and thus
that it can be removed without issue.

The feature flag to enable it was defaulting to false, and no
configuration was setting it to true.