feat(cache): [#59372] reimplement get_org_unit_path_string
This commit reimplements `get_org_unit_path_string` such that it
utilizes GraphQL's ancestor filter instead of the `org_unit_info` cache
in `LdapConverter`. This is yet another step towards eliminating state
within the integration.

The old implementation of the method is left in the code as it is used
as an implementation detail for the `get_org_unit_uuid_from_path`
method, which will be reimplemented in a future merge request.