feat(graphql): [#59372] add names GraphQL filter to org-units
This commit adds a new GraphQL filter to organisation-units making it
possible to do exact matches by name with the `names` filter.

Prior to this change a similar functionality was available via the
`query` filter, however this filter gave no guarantees with regards to
the objects returned and also had a highly flexible behavior with
regards to what was actually being searched, making it unfeasible for
entity lookups within integrations.

Using this new filter it is now possible to lookup organisation-units by
their name just as you would by their user-key or UUID, although names
do not give any guarantees with regards to uniqueness and therefore
multiple matches should be expected.

An example where this is useful; is translating an organisation unit path
to the UUID of the entity at the end of the path, for instance, imagine
the path: `"Kolding Kommune/Skole og Børn/Bjert børnehus/Administration"`.

Using a nested `names` filter, this `"Administration"` unit can be queried:
```
{
  "filter": {
    "names": ["Administration"],
    "parent": {
      "names": ["Bjert børnehus"],
      "parent": {
        "names": ["Skole og Børn"],
        "parent": {
          "names": ["Kolding Kommune"],
          "parent": null
        }
      }
    }
  }
}
```