refactor(ldap2mo): [#59372] reimplement create_org_unit recursively
This commit reimplements the `create_org_unit` method using recursion
instead of the previous implementation which used iteration.

The old implementation was difficult to understand as it used array
indexed operations to generate and modify partial and parent paths,
whereas the new implementation simply strips the incoming path one layer
at a time recursively until an existing org-unit is found or
alternatively until we reach the root org.

The new implementation seems vastly simpler to understand, and was
generated through a series of refactoring commits slowly simplifying the
old implementation till the recursive nature hidden within was revealed.