feat(init): [#59371] remove internal OS2mo-init
This commit removes the internal OS2mo-init implementation in favor of
the standard OS2mo-init container, thereby eliminating about ~600 lines
of code and aligning the integration with all other OS2mo integrations.

The motivation behind the removal, besides clearing up the codebase in
general, is that a lot more classes are found in OS2mo due to the
introduction of the FKK integration, an issue that breaks both the
OS2mo-init container and the internal implementation in this integration.
However instead of updating both the OS2mo-init container and the
somewhat equivalent implementation within this integration, it is easier
to simply convert the integration to using the standard OS2mo-init
container instead.

A regression is introduced by this change, namely an anti footgun
validator that would previously check that all values in the init
configuration was being used in the conversion mapping, however this
footgun is already present in all other integrations and as such we
gladly accept the regression when put up against the double
maintenance.