fix(ldap2mo): [#61773] avoid attempting to create details on non-existing employees
This commit fixes an issue where the integration would attempt to create
details on non-existing employees, whenever the integration got an LDAP
event UUID of an entity not found in MO, while not being configured to
create employees.

The issue was shipped in version 1.0.0 of the integration and was merged
before every incremental change to the integration was released with !101
(51e8391f), and as such have been present in all versions of the
integration until this commit.

This commit fixes the issue by introducing an early-return in the
`import_single_user` function, which stops the import whenever an
employee corresponding to the given LDPA event UUID / DN cannot be found
in MO and we were not configured to create new employees in MO.

Before this commit the code would simply generate an UUID for the new
employee, even if the integration was not configured to create it, and
proceed to figure out which fields to synchronize only to fail the
synchronization as the related employee (the randomly generated UUID)
cannot be found in MO (obviously).

The issue has not resulted in any erroneous writes, as every single
write attempt has failed, thus data integrity should be entirely
unaffected. Between version 5.1.0 (!485, 809c5f4f) and version 10.24.0
(!706, b1f3bfe0) of the integration it may have blocked synchronizations,
however it would merely have produced a backlog of messages to be
processed later. In versions prior to 5.1.0 the messages would simply
have been thrown away as any other message raising an exception.

This main affect of this commit is to reduce this backlog and to try to
get the number of constantly failing events a little more under control.