feat(api): [#61773] introduce /fixup/delete_non_existing_unique_ldap_uuids
This commit introduces a new endpoint which can be used to delete all
LDAP IT users which refer to non-existing LDAP accounts.

This situation seems common on our LDAP deployments, and while the
endpoint does not resolve the core issue, it helps us remove current
issues to get a clean slate from which we can observe, debug and fix new
occurrences.