refactor: [#59372] rework mo_to_ldap_attributes
This commit reworks the `mo_to_ldap_attributes` property, such that it
does not have a class attribute as a cache field.

The computation done to build the result is very simple and there is
absolutely no reason to cache it, and if there was, it should be done
using the `@cache` decorator rather than explicitly.