fix: [#59372] fix argument to make_ldap_object
This commit fixes an issue within the `load_ldap_cpr_object` method
calls the `make_ldap_object` function with the wrong argument, namely
passing the FastRAMQPI `context` instead of the `ldap_connection`
object.

The issue was introduced in version 10.13.0 of the integration with !672
(8f7111e9), which adjusted the parameters taken by the `make_ldap_object`
function to more accurately represented what arguments the function
actually needed, however this one call-site was missed due to lack of
typing.

This commit fixes the issue by passing the correct `ldap_connection`
object to the function instead of the incorrect `context` argument.

The issue has not resulted in any erroneous writes, thus data integrity
should be entirely unaffected. The change may however have resulted in
blocked synchronizations, such that some entities in MO have not been
correctly updated. This should have been a non-issue as the AMQP messages
should have been piling up, both serving to alert us to the issue, and
also ensuring that the backlog of broken synchronizations could be
rerun. Atleast assuming that AMQP messages are being persisted correctly.

It is unfortunate that our static analysis typing tool mypy was unable
to detect this discrepancy, however it makes sense given how much of the
code is not properly typed, in fact this very issue was detected by
introducing proper typing to the `DataLoader` class.

The issue was a subtle one that was fairly not caught in the review
process, however it should have (and likely would have) been caught if
we had had integration tests or less mocked unit-tests.
This commit does not introduce such tests, nor does it demockify the
current unit-tests, rather a future commit will address this need.