refactor: [#59372] add class attribute for settings
This commit adds a class attribute for `Settings` such that the
methods inside of `DataLoader` do not to depend on the entire
`user_context`.

Additionally, this commit refactors 11 cases where the code would extract
the `settings` directly from the context to instead using a class
attribute in which it is bound.