feat(ldap2mo): [#61773] deprecate job_function_callback in JobTitleFromADToMo
This commit deprecated the `job_function_callback` field on
`JobTitleFromADToMo` as it is a bad way to implement the default
behavior in comparison to the `default` argument on the
`get_or_create_job_function_uuid` which is used by all users of the
`JobTitleFromADToMo` custom type.

Additionally if more advanced behavior was required, it is also easy to
implement directly in the jinja template.

The field is still kept in the model for backwards compatibility, such
that the version which deprecates it works with configurations that
still have it, by just making the field a noop, such that the
configurations can be modified to remove the field, before the field
itself is removed.