feat(ldap2mo): [#61773] rework JobTitleFromADToMO
This commit reimplements the `sync_to_mo` method within
`JobTitleFromADToMO`, such that it actually does the synchronization to
MO, instead of returning a list of tasks to the caller which will
immediately await it.

The reimplementation should also improve performance as operations are
now done in parallel using `asyncio.gather` and finally it is way
shorter as the code has been vastly simplified.