fix(ldap2mo): [#61773] handle test CPR number in `import_single_user`
This commit fixes an issue where the integration would attempt to lookup
invalid testing CPR numbers in LDAP, whenever it received an event from
LDAP which has such a testing CPR number on it. This would then make the
message get retried forever as it would continuously fail CPR validations.

This commit fixes the issue by introducing a check for the common all
zeros testing CPR number before attempting to look it up in LDAP.