refactor: [#59372] make SyncTool take ldap_connection explicitly
This commit modifies SyncTool such that it takes the ldap_connection
explicitly instead of relying on extracting it from our context
'God object', this should make it clearer to the caller what is expected
of the input argument.

This is one small step in a larger refactor that tries to solve nested
cyclic dependencies within the program.