refactor: [#59372] limit CustomSpecific parameter requirements
This commit limits the `CustomerSpecific` handlers from depending on our
`context` 'God object' to instead depend on the GraphQL client which is
actually required by the implementations of the interface.

This makes it easier to limit `context` object usage in the parents as
they do not need to depend on the `context` object themselves.