fix: [#59372] fix broken AMQPSystem reference on SyncTool
This commit fixes an issue where the `DataLoader` class would attempt to
access the `AMQPSystem` on the `SyncTool` however the `AMQPSystem` on
`SyncTool` was removed in !809 (e1e1ee89).

The issue was not caught by the static type checker due to lack of typing
which in itself is due to triple-nested cyclic references in the
initialization code of the integration.

The issue was also not caught by the test-suite as it mocks literally
everything and thereby tests almost nothing.