refactor: [#59372] add lazy property for converter
This commit adds a lazy property for `LdapConverter` such that the
methods inside of `DataLoader` do not to depend on the entire
`user_context`.

This commit refactors 4 cases where the code would extract the
`converter` directly from the context to instead using a class
attribute in which it is bound.