refactor: [#59372] limit paged_search parameter requirements
This commit limits the `paged_search` function from depending on our
`context` 'God object' to instead depend on the `settings` and
`ldap_connection` and vastly smaller call interface.

This minor change however triggers ripple effects in all callers such
that their call parameters need to be adjusted as well, and even worse,
to all of the myriad of mocks pretending to be those callers and
callees.