fix: [#59372] start LDAP AMQPSystem before generating events on it
This commit fixes an issue where the `LDAPEventGenerator` would attempt
to emit messages on the LDAP AMQPSystem before it was started.
The issue arose from bad lifespan manager priorities.

The issue was introduced in version 6.25.1 of the integration with !558
(c7826887), which configured both the LDAP AMQPSystem and event
generator to start at the same time, which may have worked occasionally
depending on which service happened to be started first.

The issue was then worsened in version 6.25.2 of the integration with
!559 (b65c2c2e), which attempted to fix issues from the prior commit,
but had the unfortunate side-effect of worsening this issue while
improving another.

The core issue however was shipped in version 1.0.0 of the integration
and was merged before very incremental change to the integration was
released with !101 (51e8391f).

This commit fixes the issue by reordering the start-up order of the
services, and a future commit will attempt to make the dependency clear
by eliminating the context for sharing references between the LDAP
AMQPSystem and the `LDAPEventGenerator`.