refactor: [#59372] make LDAPEventGenerator take explicit parameters
This commit modifies `LDAPEventGenerator` such that it no longer takes our
`context` 'God object', but rather declare its dependencies directly
making it clearer to the caller what is expected of the input argument.

This is one small step in a larger refactor that tries to solve nested
cyclic dependencies within the program.