chore!: [#59372] remove engagement_uuid from from_ldap
This commit removes `engagement_uuid` from the `from_ldap` method and
thereby from the LDAP template context, such that it is no longer
possible to use it in template mappings.

This is possible as no current template mapping utilize `engagement_uuid`.

The introduction of `engagement_uuid` to the mapping context was done in
version 2.7.0 of the integration (c224ceb4, !349, #57422), which was a
part of a massive change (+657-92) attempting to introduce support for
multiple IT accounts on both import and export, however the change never
managed to become functional and reach production, and since then the
specification for how to achieve multiple IT accounts has changed
multiple times, leaving behind technical debt in the form of unused code.

BREAKING CHANGE:

It is no longer possible to use `engagement_uuid` within LDAP template
mappings. It is our belief that noone uses this functionality as thus
that noone should be affected by this breaking change.