ci(deptry): [#59372] stop ignore client dependencies
This commit removes the filter on autogen-client dependencies when being
checked by deptry sometimes leading to errors about unused packages that
are in fact used by our autogen-client.

We originally ignored the autogen-client as we did not want to install
the websockets package that it optionally depends on, but it is probably
better to just install all references packages so we do not accidentally
uninstall a package that is used by our autogen-client, simply because
it is not used in the rest of the project.

In short, if the autogen-client depends on a package we do in fact want
that package installed as our autogen-client might not work otherwise.