fix: [#59372] ensure load_ldap_OUs works on standard LDAP
This commit fixes an issue where the `load_ldap_OUs` function would
always return `False` for the `empty` field on standard LDAP, as the
code has been written to exclusively work on Microsofts non-standard
compliant LDAP implementation Active Directory.

The issue specificially arose from trying to lookup users by using the
Microsoft non-standard `user` `objectClass` instead of using the
standard `inetOrgPerson` or similar.

The issue was shipped in version 2.1.0 of the integration with !343
(4a0a35e4), back when the integration did not have a local development
environment running on standard LDAP and back when we did not care to
support standard LDAP.

This commit resolves the issue by introducing a configuration variable
to support different `objectClass`es for selecting users, defaulting to
`inetOrgPerson` on standard LDAP and to `user` on Microsofts non-standard
implementation.