test: [#59372] convert test_load_ldap_OUs to use ldap3 mocking
This commit changes the `test_load_ldap_OUs` test function from using
standard `unittest.mock` mocks attempting to imitate how an LDAP server
works to instead using the mocking / testing functionality provided with
our LDAP library.

This change was made as a future restructuring of the code would make
our test mocks fail in a flaky way due to parallelization that was not
taken into account in our mocking.

In general we much prefer to use proper mocking libraries implementing
the protocol they are mocking, such as `respx` for `httpx` over trying
to implement the entire HTTP protocol ourselves using `unittest.mock` or
foregoing that and instead making extremely fragile test code as the
mocks do not actually correspond to the real behavior of the code, thus
reducing our test-suite to testing nothing.