feat(graphql)!: [#60458] move inherit flag to manager resolver
BREAKING CHANGE:

The filter for managers on organizational units have changed its type
from `ManagerFilter` to `OrgUnitsboundmanagerfilter`, indicating that
`org_units` can no longer be set on the filter as it is seeded from the
context. Setting it prior to this change would not work either, as it
was always overridden inside the handler.

To migrate simply rename `ManagerFilter` to
`OrgUnitsboundmanagerfilter`.