chore: [#59372] fix flakiness in the test-suite
This commit fixes flakiness in the `test_application.py::test_ldap2mo`
test (and potentially others), where it would occasionally produce the error:
```
sqlalchemy.exc.ProgrammingError: (psycopg.ProgrammingError)
Explicit rollback() forbidden within a Transaction context.
(Either raise Rollback() or allow an exception to propagate out of the context.)
```
This error would be emitted from the `LDAPEventGenerator` whenever an
`asyncio.CancellationError` was triggered within the SQLAlchemy code
updating the `LastRun` database table.

The `asyncio.CancellationError` would be triggered by the shutdown of
the `LDAPEventGenerator` itself triggered by the shutdown of the
lifespan on the `FastAPI` lifespan, triggered by the shutdown of the
`TestClient` test-fixture.

The issue itself seems to spawn from improper `asyncio.CancellationError`
handling within SQLAlchemy, however to resolve the issue we have opted
to simply wrap our SQLAlchemy code in an `asyncio.shield` to protect it
from `asyncio.CancellationError`s, such that cancellation is delayed
until after the current iteration is done.