feat(ldap2mo): [#61915] set default validity in ldap2mo
This commit sets the default validity for ldap2mo mapping objects to
`[now, infinity[`, such that one does not have to provide a validity for
these objects.

The long term goal is changing the behavior of explicitly validity to
mean that we wish to assert temporal data, instead of just writing
actual state data, but for now this change is simply made so the
deployments can be adapted in a backwards compatible way before
disallowing setting validity explicitly.