feat(ldap2mo)!: [#61915] disallow validity in ldap2mo
This commit disallows setting validity for the ldap2mo mapping objects,
such that it is now only possible get validity set using the default.

In the future a functionality to add temporal assertions may be added,
reintroducing the validity argument, but for now it is disallowed.

BREAKING CHANGE:

It is no longer possible to explicitly set the validity setting in
ldap2mo mappings.