feat(ldap2mo): [#61915] use injector to create injection
This commit implements the newly added `_injector_` field on `ldap_to_mo`
mappings to create injections between objects in LDAP and objects in MO.

It is therefore now possible to use Jinja templates to decide which
elements in LDAP and MO map to one another, simply by ensuring that the
Jinja template returns the same value for objects in the same injection.
The `_injector_` template has the same template context as all other
templates, except that `ldap` object is not available, rather an `obj`
field is set in the context to the loaded MO object and the converted
LDAP object respectively.

The behavior whenever the `_injector_` field is not set remains
unchanged.