feat: [#61915] relax address_type and itsystem template validations
This commit relaxes the requirements for the template validations on
address_type for addresses and itsystem for it-users, such that any
template can be used, rather than just the one specified in the pydantic
validator.

The pydantic validation was originally introduced version 2.6.1 of the
integration (f6a5f8ba, !365), back when all customer configuration was
present in the repository. The validation was converted from a test-case
doing the same validation, which was introduced in version 1.16.0 of the
integration (80d3057e, !190).

While researching the original introduction of the code, it did not seem
clear why it was introduced as an invariant other than to be an
anti-footgun turned footgun, and thus Chesterton's Fence has been taken
down.