feat: [#61915] add the LDAP AMQPSystem to context whether started or not
This commit ensures that the LDAP AMQPSystem is always added to the
FastRAMQPI context whether it is getting started or not, i.e. whether we
wish to listen to its messages or not.

This is required as we might still want to emit messages, or may
reference the LDAP AMQPSystem, even if we do not actively listen to it.

In the future the system will be removed entirely in favor of HTTP
events, but for now, this change is necessary for running with
`listen_to_changes_in_ldap: False`.