refactor: [#59372] move template.globals to environment.globals
This commit moves our `template.globals` to `environment.globals` such
that the environment exposes all the functionality needed in templates
instead of having to dynamically expand all templates with the exact
same globals.

This aligns globals with our filters, which are indeed handled on the
environment as the jinja library suggests it, thus bringing us closer to
how the jinja library is intended to be used.