refactor: [#59372] utilize standard UUID class to remove brackets
This commit utilizes UUIDs ability to strip curly braces instead of the
template function to ensure that the template functions are not used
from the rest of the code base in order to promote separation of
concerns and ease refactoring in the future. Additionally we prefer to
use the standard library for standard tasks, rather than using our own
home-rolled version of things. Finally the function is named poorly, as
it should be curly braces not brackets.