feat(api): [#59372] remove /LDAP/Employee post endpoint
This commit removes the `/LDAP/{json_key}` HTTP POST endpoint, which
could be used to write directly to LDAP bypassing the entire
integration, including all failsafes and guard-rails.

If we wish to do so in the future, one can simple `exec` into the
container, and run a python intepreter importing the relevant functions
instead. This is obviously more difficult to do, but rightly so, as it
should never be a standard operation.