chore: [#59372] remove check_ldap_to_mo_references
This commit removes the `check_ldap_to_mo_references` configuration
validator, as it was implemented in a non-robust way, checking
specifically for `ldap.` substrings in the template, which would
immediately break assuming someone wrote `getattr(ldap, fieldname)`
instead.

We do not want non-robust validation, as it creates a false sense of
security, so we rather prefer to just fail on runtime when the actual
ldap object is queried.

The validation may be reintroduced in the future if we deem it
necessary, but for now we prefer to simplify the code and reduce the
complexity.