chore: [#59372] remove clean_get_current_method_from_template_string
This commit removes `clean_get_current_method_from_template_string`
from the configuration validation, which was previously used to do some
kind of weird special case cleanup on the template strings used in the
`mo_to_ldap` mapping, by running a regex on the non-regular jinja
template code.

The cleanup routine was implemented in an extremely fragile way and could
easily result in false positives if one happened to have a `")"` anywhere
in the jinja template string.

Additionally the test for the functionality was severely broken not even
operating on a syntactically valid jinja template, and resulting in even
further invalid jinja after processing.

We do not want non-robust validation, as it creates a false sense of
security, so we rather prefer to just fail on runtime when the actual
code is run.

The validation may be reintroduced in the future if we deem it
necessary, but for now we prefer to simplify the code and reduce the
complexity.