chore: [#59372] remove check_ldap_attributes
This commit removes the `check_ldap_attributes` configuration validator,
as it was implemented in a non-robust way, checking specifically for
`ldap.` substrings in the template, which would immediately break
assuming someone wrote `getattr(ldap, fieldname)` instead, but also
depending on `json_key`s being it-system user-keys or address-type
user-keys, besides the fact that it also hard-coded template string
values to look for.

The validator was intended to check that mapping multiple MO
values into a single LDAP field would work sanely, by ensuring the
LDAP field in such situation was multi-valued.

Fortunately this check was made and the corresponding behavior was made
obsolete by a restructuring in the mo2ldap synchronization code, which
now only maps a single MO object to LDAP, using the MO object with the
lowest UUID if multiple matches are found.
As such the use for the validator has shrunk severely.

Now in theory multiple different mappings could be writing to the same
LDAP field and thus need to be combined in a multivalued field, a
situation which actually happens in our production configuration, however
the validator did not check for this situation instead checking the
different mappings in isolation, so protection against this scenario was
not provided by the validation, even if it would seem as if it was.

We do not want non-robust validation, as it creates a false sense of
security (such as assuming it protects across multiple different
mappings), so we rather prefer to just fail on runtime when the actual
code is run.

The validation may be reintroduced in the future if we deem it
necessary, but for now we prefer to simplify the code and reduce the
complexity. Additionally we also have a plan to vastly simplify the
mo2ldap mapping such that a single mapping exists instead of multiple
individual mappings.