feat: [#59372] add feature flag for cpr field fiddling
This commit adds a feature flag to turn off the cpr field fiddling
logic, which attempts to inspect jinja templates to find one that
uniquely points to the field in LDAP that stores the CPR number.

We much prefer a simple static configuration in which the configuration
has a hard-coded value for the LDAP field that contains the cpr number,
instead of having to do this magic template fiddling logic, and thus such
a configuration value is introduced, and the fiddling logic will be
removed in a future merge request.