feat!: [#59372] remove it-system user-key fiddling
This commit removes the ability to configure the MO LDAP UUID it-system
user-key using a mapping jinja-template, by removing the
`find_ldap_it_system` function and its corresponding functionality, such
that the it-system user-key must instead be explicitly specified with the
`LDAP_IT_SYSTEM` environmental variable.

The removal has multiple motivation, for instance that having to
configure an unused jinja-template mapping simply to set the it-system
user-key seems overly complex compared to setting it directly.
Additionally the implementation of the it-system user-key detector was
not very robust, simply templating a broken context with a single random
value "detection-key" into all `user-key` templates in the entire
`ldap_to_mo` mapping, trying to find the "detection-key" in the output,
which would immediately break if the template did not output the input
value completely unmodified.

We do not want overly complex configuration, and we do not want to use
non-robust heuristics for configuring essential parts of our integration.

BREAKING CHANGE:

It is no longer possible to configure the MO LDAP UUID it-system user-key
indirectly using reverse mapping jinja-templates, instead the
configuration must be done directly with the `LDAP_IT_SYSTEM`
environmental variable.