feat!: [#59372] fail hard on bad template configuration
This commit changes the configuration of the Jinja template environments
with regards to how to deal with undefined behavior. Specifically the
configuration has been changed from `Undefined` to `StrictUndefined`.

With `Undefined` converting to an undefined entity to a string results in
the empty string, while converting to a boolean returns `False`, etc.
With `StrictUndefined` an exception is raised for these cases instead.

The goal of the change is to avoid falling back on unexpected behavior
if the template context does not match what the template writer expects,
and thus to force the template writer to handle situations of fields not
being set, or misconfigurations explictly instead of these being silently
coerced into falsy values.

For more details, see:

* https://jinja.palletsprojects.com/en/3.1.x/api/#undefined-types

BREAKING CHANGE:

The templating system will now raise exceptions on undefined behavior
instead of silently returning falsy values.