test(mo2ldap): [#59372] remove CN from employee test
This commit removes mapping to, and asserting the state of the `CN`
'attribute' as it is not possible to set this value using the
`modify` function of the `ldap3` library, which is currently used by
the integration for all mapped values.

This is because `CN` is not actually an attribute, but rather a part
of the `DN` and thus must be set using the specialized `modify_dn`
function of the `ldap3` library. The same seems to apply for `OU`, `O`
and others.

See: https://ldap3.readthedocs.io/en/latest/modifydn.html