fix(mo2ldap): [#59372] avoid failing silently on bad writes
This commit changes the error handling within the `modify_ldap_object`
method, such that it now raises a `RequeueMessage` exception on write
errors instead of just emitting a warning and otherwise pretending like
nothing bad happened.

This commit may thus cause a lot of messages to get stuck in the queues
as they were previously failing silently, but it is a change that must
be made to ensure that messages are not just dropped whenever a write
fails for whatever reason, be it temporary downtime, misconfiguration,
etc.