feat(mo2ldap)!: [#59372] remove mo_to_ldap mapping
This commit removes the `mo_to_ldap` mapping from the configuration of
the integration, such that it is no longer possible to set the old
configuration.

This commit also remove the code utilizing the the old configuration to
synchronize from mo to ldap, but does not garbage collect all of the now
dead functions, rather leaving that cleanup for the following commits,
in an attempt to keep this changeset as minimal as possible.

BREAKING CHANGE:

It is no longer possible to set `mo_to_ldap` on the configuration,
instead `mo2ldap` must be set with a single template for all mappings.