feat(ldap2mo?)!: [#59372] remove manual import triggering
This commit removes the triggering of manual imports from LDAP to MO
that happens automatically from the MO to LDAP synchronization.

The removal happens as no customer actually has `manual_import_only`
configured, and as the logic is very, very hard to follow and attempts
at analysing the git-history to avoid pulling down Chesterton's fence
has come up fruitless.

In the end we do not believe that processes specifically configured to
be manual should happen automatically and thus the removal has been
done.

BREAKING CHANGE:

If you were relying on manual imports to be done automatically this may
be a breaking change for you. The solution is probably to do manual
imports manually.