refactor: [#59372] remove employee refresh on new LDAP user
This commit removes the code that triggers an employee refresh event in
MO whenever a new user in LDAP has been created.

The refresh was triggered from within the MO to LDAP synchronization
code, and would effectively just retrigger the code later, however no
such retriggering should be required as we are currently running the
synchronization and will synchronize everything.

Additionally, if we are using IT-users to map LDAP and MO accounts, an
event will be emitted for the newly created IT-user which will
eventually trigger an employee refresh event anyway.

Attempts at analysing the git-history to avoid pulling down Chesterton's
fence has come up fruitless, but it seems likely that the code is from a
bygone era of `uuids_to_ignore` and partial LDAP updates using
object-specific events and mappings.