fix: [#62597] ensure AssertionError is raised correctly
This commit fixes a bug in the usernames generator, where an
`AssertionError` was asserted instead of raised.

This was likely due to a typo, but was unfortunately not detected by any
static analysis tool. We must race `AssertionError`s instead of `assert
False` as the latter can be optimized away using the `-O` flag to the
python interpreter.