refactor: [#62597] split generate_username out
This commit splits out `generate_username` from `generate_dn` such that
it becomes a more freestanding function that can be called from our
future jinja global.

The change is a little more involved than what might first be expected,
because a lot of code has been duplicated between the main class and its
subclass, and since the `generate_username` was already present in the
subclass.