feat: [#62597] rework exception handling to log stacktraces
This commit changes how exceptions are logged from within the
`*_reject_on_failure` decorators, such that stacktraces are produced
using `logger.exception` instead of avoiding them using `logger.warning`
and friends.

Some of the log levels have also been changed, for instance because it
was considered silly to log a warning on a message requeue which may
happen for lots of valid reasons.