feat(mo2ldap): [#59372] remove special-case create code
This commit removes the hardcoded `sAMAccountName` mapping from the
username generator, such that it must instead be set using the `mo2ldap`
jinja template.

This also applies to `userPrincipalName` in Alleroed, but the change
essentially has the same goal, we do not want hardcoded behavior when it
is just a easy to configure the behavior using the `mo2ldap` jinja
template.